requirements.txt
asgiref==3.2.7
cffi==1.14.0
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
fastpbkdf2==0.2
fuzzy-extractor==0.3
numpy==1.18.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1